+++ pkg/kiosk/grafana_genericoauth_login.go
@@ -36,7 +36,11 @@ func GrafanaKioskGenericOauth(cfg *Config, messages chan string) {
 		panic(err)
 	}
 
-	var generatedURL = GenerateURL(cfg.Target.URL, cfg.General.Mode, cfg.General.AutoFit, cfg.Target.IsPlayList)
+	var generatedURL = GenerateURL(
+		cfg.Target.URL,
+		cfg.GrafanaOptions.KioskMode,
+		cfg.GrafanaOptions.AutoFit,
+		cfg.Target.IsPlayList)
 
 	log.Println("Navigating to ", generatedURL)